+++ src/vi/k2600/k2600.mstest/Tsp/TspScriptTests.cs
@@ -2,10 +2,11 @@ using System;
 using System.Diagnostics;
 using System.IO;
 using System.Text;
+using Microsoft.Extensions.Logging;
 using cc.isr.Std.Tests;
 using cc.isr.Std.Tests.Extensions;
+using cc.isr.VI.Tsp.Script.ExportExtensions;
 using cc.isr.VI.Tsp.Script.SessionBaseExtensions;
-using Microsoft.Extensions.Logging;
 
 namespace cc.isr.VI.Tsp.K2600.MSTest.Tsp;